# HTTP: Optimise response rendering and response parsing