… akka-bench-jmh/src/main/scala/akka/http/HttpBlueprintBenchmark.scala
… akka-bench-jmh/src/main/scala/akka/http/HttpRequestParsingBenchmark.scala
… akka-bench-jmh/src/main/scala/akka/http/HttpResponseRenderingBenchmark.scala
… akka-http-core/src/main/scala/akka/http/scaladsl/model/HttpHeader.scala
… akka-http-core/src/test/scala/akka/http/impl/engine/ws/WebSocketIntegrationSpec.scala
… akka-http-testkit/src/main/scala/akka/http/scaladsl/testkit/RouteTest.scala
… akka-parsing/src/main/scala/akka/parboiled2/ParserInput.scala