@ktoso 2016-07-26
Other lessons learnt:
- fully caching the Date header more than we do now yields around 100k more rendering, so not as much as one might think.
- caching the content type rendered values was a nice boost, there's potential in this trick. though hard to apply to other dynamic things.