@ktoso 2016-07-28
This should now pass, please review @drewhk @johanandren