@ktoso 2016-07-28
Resolves https://github.com/akka/akka/issues/21062